change int to double for recipeId

+++ src/main/java/com/mgilhespy/tastycreations/models/Like.java
@@ -24,7 +24,7 @@ public class Like {
     private Long id;
 
     @NotNull
-    private int recipeId;
+    private double recipeId;
 
     @Column(updatable=false)
     @DateTimeFormat(pattern="yyyy-MM-dd")
